// ps-frontend/src/model/template.ts
export interface ConditionT {
  field: string;
  operator: 'eq' | 'neq' | 'lt' | 'le' | 'gt' | 'ge';
  value: string;
}